backend/requirements.txt
python-jobspy
playwright
fastapi
uvicorn
pymongo
scikit-learn
httpx
pytest
pandas
python-dotenv
mongomock
supabase
email-validator
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymupdf
pytesseract
pillow
python-docx
docx2txt
groq
python-dateutil
